requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Werkzeug==0.9.4
itsdangerous==0.23
python-firebase==1.1
requests==2.0.1
wsgiref==0.1.2
